Converted CSGroups into a class
